move AbstractFileClassLoader into dotty.tools.io
as per Dale's comment
https://github.com/lampepfl/dotty/pull/18536#discussion_r1324487683
