+++ compiler/src/dotty/tools/io/AbstractFileClassLoader.scala
@@ -11,6 +11,7 @@
  */
 
 package dotty.tools
+package io
 
 import scala.language.unsafeNulls
