+++ scaladoc/src/dotty/tools/scaladoc/snippets/SnippetCompiler.scala
@@ -12,8 +12,7 @@ import dotty.tools.dotc.interfaces.{SourceFile => ISourceFile}
 import dotty.tools.dotc.reporting.{ Diagnostic, StoreReporter }
 import dotty.tools.dotc.parsing.Parsers.Parser
 import dotty.tools.dotc.{ Compiler, Run }
-import dotty.tools.AbstractFileClassLoader
-import dotty.tools.io.{AbstractFile, VirtualDirectory}
+import dotty.tools.io.{AbstractFile, AbstractFileClassLoader, VirtualDirectory}
 import dotty.tools.dotc.util.Spans._
 import dotty.tools.dotc.interfaces.Diagnostic._
 import dotty.tools.dotc.util.{ SourcePosition, NoSourcePosition, SourceFile, NoSource }
