+++ staging/src/scala/quoted/staging/QuoteDriver.scala
@@ -5,8 +5,7 @@ import dotty.tools.dotc.ast.tpd
 import dotty.tools.dotc.Driver
 import dotty.tools.dotc.core.Contexts.{Context, ContextBase, FreshContext}
 import dotty.tools.dotc.quoted.QuotesCache
-import dotty.tools.io.{AbstractFile, Directory, PlainDirectory, VirtualDirectory}
-import dotty.tools.AbstractFileClassLoader
+import dotty.tools.io.{AbstractFile, AbstractFileClassLoader, Directory, PlainDirectory, VirtualDirectory}
 import dotty.tools.dotc.reporting._
 import dotty.tools.dotc.util.ClasspathFromClassloader
 import scala.quoted._
